// frontend/plugins/welcome/src/api/models/EntClubappStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntClubappStatusEdges,
    EntClubappStatusEdgesFromJSON,
    EntClubappStatusEdgesFromJSONTyped,
    EntClubappStatusEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntClubappStatus
 */
export interface EntClubappStatus {
    /**
     * Clubstatus holds the value of the "clubstatus" field.
     * @type {string}
     * @memberof EntClubappStatus
     */
    clubstatus?: string;
    /**
     * 
     * @type {EntClubappStatusEdges}
     * @memberof EntClubappStatus
     */
    edges?: EntClubappStatusEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntClubappStatus
     */
    id?: number;
}

export function EntClubappStatusFromJSON(json: any): EntClubappStatus {
    return EntClubappStatusFromJSONTyped(json, false);
}

export function EntClubappStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntClubappStatus {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'clubstatus': !exists(json, 'clubstatus') ? undefined : json['clubstatus'],
        'edges': !exists(json, 'edges') ? undefined : EntClubappStatusEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntClubappStatusToJSON(value?: EntClubappStatus | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'clubstatus': value.clubstatus,
        'edges': EntClubappStatusEdgesToJSON(value.edges),
        'id': value.id,
    };
}